refactor: move utils.go to /utils
