+++ driver_infrastructure/random_host_selector.go
@@ -19,13 +19,14 @@ package driver_infrastructure
 import (
 	"awssql/error_util"
 	"awssql/host_info_util"
+	"awssql/utils"
 	"math/rand"
 )
 
 type RandomHostSelector struct{}
 
 func (r *RandomHostSelector) GetHost(hosts []host_info_util.HostInfo, role host_info_util.HostRole, props map[string]string) (host_info_util.HostInfo, error) {
-	eligibleHosts := FilterSlice(hosts, func(hostInfo host_info_util.HostInfo) bool {
+	eligibleHosts := utils.FilterSlice(hosts, func(hostInfo host_info_util.HostInfo) bool {
 		return role == hostInfo.Role && hostInfo.Availability == host_info_util.AVAILABLE
 	})
